ifscsdr_devices_controller/requirements.txt
pyudev
pyrtlsdr
